fix(invites): split claim helpers so route bundle drops cloudflare:workers
Vite couldn't bundle the claim/decline routes because both transitively
imported `getDb` (which imports `cloudflare:workers`) through
`server/competition-invites/claim.ts`. The dev server saw the unresolvable
import in the client graph and crashed every request, taking the e2e
suite with it.

Move the pure helpers (`identityMatch`, `assertInviteClaimable`, types,
`InviteNotClaimableError`) into a sibling `identity.ts` and re-export from
`claim.ts` for back-compat. Routes now import from `identity.ts` only.

Also drops the useless `case "not_found":` fallthrough flagged by biome —
`default` already covers it.
