+++ apps/wodsmith-start/src/server/competition-invites/claim.ts
@@ -16,6 +16,11 @@
  * hop happens inside the Stripe webhook workflow (Phase 2 sub-arc C). This
  * module is side-effect-free beyond DB reads, so it's safe to call from
  * route loaders, server functions, and tests.
+ *
+ * Pure helpers (`identityMatch`, `assertInviteClaimable`, types) live in
+ * [[./identity.ts]] so route files can import them without dragging in
+ * `getDb` (and the `cloudflare:workers` binding) — Vite chokes on the latter
+ * when bundling the client.
  */
 // @lat: [[competition-invites#Claim resolution]]
 
@@ -23,51 +28,26 @@ import { and, eq } from "drizzle-orm"
 import { getDb } from "@/db"
 import {
   COMPETITION_INVITE_ACTIVE_MARKER,
-  COMPETITION_INVITE_STATUS,
   type CompetitionInvite,
-  type CompetitionInviteStatus,
   competitionInvitesTable,
 } from "@/db/schemas/competition-invites"
 import { hashInviteClaimToken } from "@/lib/competition-invites/tokens"
-import { normalizeInviteEmail } from "./issue"
+import {
+  assertInviteClaimable,
+  InviteNotClaimableError,
+} from "./identity"
 
-// ============================================================================
-// Types
-// ============================================================================
-
-export type InviteClaimableError =
-  | "not_found"
-  | "expired"
-  | "declined"
-  | "revoked"
-  | "already_paid"
-
-export type IdentityMatchResult =
-  | { ok: true }
-  | { ok: false; reason: "wrong_account" | "needs_sign_in" | "needs_sign_up" }
-
-export interface IdentityMatchSession {
-  /** `null` / `undefined` when the visitor is signed out. */
-  email?: string | null
-}
-
-export interface IdentityMatchOptions {
-  /**
-   * Whether WODsmith has a user account matching `invite.email`. The route
-   * loader supplies this by querying `userTable` — it isn't tracked on the
-   * invite row because a user may sign up *later*.
-   */
-  accountExistsForInviteEmail: boolean
-}
-
-export class InviteNotClaimableError extends Error {
-  readonly reason: InviteClaimableError
-  constructor(reason: InviteClaimableError, message?: string) {
-    super(message ?? `Invite is not claimable: ${reason}`)
-    this.name = "InviteNotClaimableError"
-    this.reason = reason
-  }
-}
+// Re-export the pure helpers/types so existing imports of this module keep
+// working. Route files should import from `./identity` directly.
+export {
+  assertInviteClaimable,
+  identityMatch,
+  InviteNotClaimableError,
+  type IdentityMatchOptions,
+  type IdentityMatchResult,
+  type IdentityMatchSession,
+  type InviteClaimableError,
+} from "./identity"
 
 // ============================================================================
 // Resolve by token
@@ -97,91 +77,6 @@ export async function resolveInviteByToken(
     .where(eq(competitionInvitesTable.claimTokenHash, hash))
     .limit(1)
   return rows[0] ?? null
-}
-
-// ============================================================================
-// Claimability
-// ============================================================================
-
-/**
- * Throws {@link InviteNotClaimableError} unless the invite is:
- * - status = `pending`,
- * - has a live `claimTokenHash` and `expiresAt` in the future.
- *
- * `accepted_paid`, `declined`, `expired`, and `revoked` each map to a
- * distinct `reason` so the route can render a tailored message (and so
- * support queries can tell them apart from a generic "invalid").
- */
-export function assertInviteClaimable(
-  invite: CompetitionInvite,
-  now: Date = new Date(),
-): void {
-  switch (invite.status as CompetitionInviteStatus) {
-    case COMPETITION_INVITE_STATUS.PENDING:
-      break
-    case COMPETITION_INVITE_STATUS.ACCEPTED_PAID:
-      throw new InviteNotClaimableError("already_paid")
-    case COMPETITION_INVITE_STATUS.DECLINED:
-      throw new InviteNotClaimableError("declined")
-    case COMPETITION_INVITE_STATUS.EXPIRED:
-      throw new InviteNotClaimableError("expired")
-    case COMPETITION_INVITE_STATUS.REVOKED:
-      throw new InviteNotClaimableError("revoked")
-    default:
-      throw new InviteNotClaimableError("not_found")
-  }
-
-  // Token must be live. The lookup already required a non-null hash, but
-  // guard against a race (e.g. webhook nulled the hash between resolve and
-  // assert) so we never claim a just-terminated invite.
-  if (!invite.claimTokenHash) {
-    throw new InviteNotClaimableError("not_found")
-  }
-
-  if (invite.expiresAt && now >= invite.expiresAt) {
-    throw new InviteNotClaimableError("expired")
-  }
-
-  if (invite.activeMarker !== COMPETITION_INVITE_ACTIVE_MARKER) {
-    throw new InviteNotClaimableError("not_found")
-  }
-}
-
-// ============================================================================
-// Identity match
-// ============================================================================
-
-/**
- * Decide what happens next based on the visitor's session and whether a
- * WODsmith account exists for the invite's locked email.
- *
- * - signed in, session email matches invite email → `{ ok: true }`
- * - signed in, session email differs               → `wrong_account`
- * - signed out, account exists                     → `needs_sign_in`
- * - signed out, no account                         → `needs_sign_up`
- *
- * Case-insensitive on both sides via {@link normalizeInviteEmail}, so
- * "Mike@Example.com" and "mike@example.com" resolve identically.
- */
-export function identityMatch(
-  session: IdentityMatchSession | null | undefined,
-  invite: CompetitionInvite,
-  opts: IdentityMatchOptions,
-): IdentityMatchResult {
-  const sessionEmail = session?.email
-  const inviteEmail = normalizeInviteEmail(invite.email)
-
-  if (sessionEmail) {
-    if (normalizeInviteEmail(sessionEmail) === inviteEmail) {
-      return { ok: true }
-    }
-    return { ok: false, reason: "wrong_account" }
-  }
-
-  if (opts.accountExistsForInviteEmail) {
-    return { ok: false, reason: "needs_sign_in" }
-  }
-  return { ok: false, reason: "needs_sign_up" }
 }
 
 // ============================================================================
