+++ lat.md/competition-invites.md
@@ -84,11 +84,13 @@ The line parser auto-detects delimiter per line — tab wins when present so a G
 
 ## Claim resolution
 
-Read-side helpers for the `/compete/$slug/claim/$token` route live in [[apps/wodsmith-start/src/server/competition-invites/claim.ts]]. The module is side-effect-free beyond DB reads so the claim route loader, the decline server fn, and any future integration test can all reuse it.
+Read-side helpers for the `/compete/$slug/claim/$token` route live in [[apps/wodsmith-start/src/server/competition-invites/claim.ts]] (DB reads) and [[apps/wodsmith-start/src/server/competition-invites/identity.ts]] (pure helpers route files can import without dragging `getDb` into the client bundle).
 
-[[apps/wodsmith-start/src/server/competition-invites/claim.ts#resolveInviteByToken]] hashes the plaintext token via [[apps/wodsmith-start/src/lib/competition-invites/tokens.ts#hashInviteClaimToken]] and looks up the row on `claimTokenHash`. It returns `null` on miss so the route can render the generic "invalid link" page without leaking whether a hash ever existed. [[apps/wodsmith-start/src/server/competition-invites/claim.ts#assertInviteClaimable]] throws [[apps/wodsmith-start/src/server/competition-invites/claim.ts#InviteNotClaimableError]] when status, `claimTokenHash`, `expiresAt`, or `activeMarker` fail the live checks — each terminal state maps to a distinct `reason` (`expired`, `declined`, `revoked`, `already_paid`, `not_found`) so the route can render tailored copy.
+The split exists because Vite chokes on `cloudflare:workers` when bundling client code, and `getDb` transitively imports it. Route files import only from `identity.ts`; server-side callers can use either module.
 
-[[apps/wodsmith-start/src/server/competition-invites/claim.ts#identityMatch]] is the pure email-lock gate. Given `{ email }` from the session (or `null`) plus the invite plus an `accountExistsForInviteEmail` boolean (resolved by the caller against `userTable`), it returns the discriminated result the route loader branches on: `{ ok: true }`, `{ ok: false, reason: "wrong_account" | "needs_sign_in" | "needs_sign_up" }`. Case-insensitive on both sides via [[apps/wodsmith-start/src/server/competition-invites/issue.ts#normalizeInviteEmail]] so `Mike@Example.com` and `mike@example.com` compare equal.
+[[apps/wodsmith-start/src/server/competition-invites/claim.ts#resolveInviteByToken]] hashes the plaintext token via [[apps/wodsmith-start/src/lib/competition-invites/tokens.ts#hashInviteClaimToken]] and looks up the row on `claimTokenHash`. It returns `null` on miss so the route can render the generic "invalid link" page without leaking whether a hash ever existed. [[apps/wodsmith-start/src/server/competition-invites/identity.ts#assertInviteClaimable]] throws [[apps/wodsmith-start/src/server/competition-invites/identity.ts#InviteNotClaimableError]] when status, `claimTokenHash`, `expiresAt`, or `activeMarker` fail the live checks — each terminal state maps to a distinct `reason` (`expired`, `declined`, `revoked`, `already_paid`, `not_found`) so the route can render tailored copy.
+
+[[apps/wodsmith-start/src/server/competition-invites/identity.ts#identityMatch]] is the pure email-lock gate. Given `{ email }` from the session (or `null`) plus the invite plus an `accountExistsForInviteEmail` boolean (resolved by the caller against `userTable`), it returns the discriminated result the route loader branches on: `{ ok: true }`, `{ ok: false, reason: "wrong_account" | "needs_sign_in" | "needs_sign_up" }`. Case-insensitive on both sides so `Mike@Example.com` and `mike@example.com` compare equal.
 
 ## Claim routes
